(Ambiance/gtk3) Move the states rules at the end of the file.
Elements with both a class and a pseudoclass were being assigned
properties belonging only to the class, because the CSS precedence
was the same for both, and in these cases the last one wins.

Example:

*:selected
{ background-color: @selected_bg_color; }

.cell 
{ background-color: @base_color; }

With these two rules, a .cell:selected element would get the
@base_color instead of the @selected_bg_color.

The fix for this is to post the selector rules at the end
of the file. This may fix many things, but it particularly
fixes Banshee's ListView custom widget.