Ambiance: remove redundant attributes
Before revision 311 [1], the attributes for states pseudoclasses were
not being applied properly (as by the rules of CSS precedence, these
had the same specificity as other class rules, so the last one to be
specified won), so then these attributes were being hardcoded in a
duplicated manner in many rules with pseudoclasses.

Such was the case for the following attributes:

row:selected {
    color: @selected_fg_color;
}
row:selected:backdrop {
    color: @backdrop_selected_fg_color;
}
.entry:selected,
.entry:selected:focused {
    color: @selected_fg_color;
}
GtkIconView.view.cell:selected,
GtkIconView.view.cell:selected:focused {
    color: @selected_fg_color;
}

But, as revision 311 [1] added the pseudoclasses at the end of the
CSS file, then these attributes finally took effect in all cases,
and therefore they become redundant, so it is safe to remove them.

[1] http://bazaar.launchpad.net/~ubuntu-art-pkg/ubuntu-themes/trunk/revision/311