gnome-terminal: fix scrollbar background
This is a fix specifically for gnome-terminal, because scrollbars should have
the same trough color as the background of the terminal itself (instead of the
theme's color). Thus, change the background for those to transparent.

We can't do that for all scrollbars, as for example WebKit views will not draw
correctly behind transparent scrollbars.